% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCommittees.R
\name{getCommittees}
\alias{getCommittees}
\title{Get legislative committees}
\usage{
getCommittees(biennium, type = c("df", "list", "xml"))
}
\arguments{
\item{biennium}{Character vector representing the biennium(s) to be
searched. Each argument should take the form "XXXX-YY"}

\item{type}{One of "df", "list", or "xml". Specifies the format for
the output.}
}
\value{
\code{getCommittees} returns an object of type equal to the
\code{type} argument (defaults to dataframe)
}
\description{
Get a list of all committees that were active during the biennium,
along with their respective committee code
}
\examples{
getCommittees("2007-08")
}
