% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSponsors.R
\name{getSponsors}
\alias{getSponsors}
\title{Get legislators}
\usage{
getSponsors(biennium, type = c("df", "list", "xml"))
}
\arguments{
\item{biennium}{Character vector representing the biennium(s) to be
searched. Each argument should take the form "XXXX-YY"}

\item{type}{One of "df", "list", or "xml". Specifies the format for
the output.}
}
\value{
\code{getSponsors} returns an object of type equal to the
\code{type} argument (defaults to dataframe)
}
\description{
Get a list of all sponsors (all congressmembers) for a given biennium
}
\examples{
getSponsors("2007-08")
}
