% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/washex.r
\docType{package}
\name{washex-package}
\alias{washex}
\alias{washex-package}
\title{washex: Washington State Legislative Explorer}
\description{
Gets data from the Washington State Legislature.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rwrandles/washex-r}
  \item Report bugs at \url{https://github.com/rwrandles/washex-r/issues}
}

}
\author{
\strong{Maintainer}: Rohnin Randles \email{rrandles@princeton.edu}

}
\keyword{internal}
